[DO NOT MERGE] Find files related to a Submission
All of the File resources that reference the supplied Submission
resource will be retrieved from the Fedora server and included in the
deposit service data model that is being constructed.  Sample data is
updated to contain three File resources, only two of which reference the
data's Submission.

Unit and Integration tests were updated to confirm that the appropriate
number of File resources were found.  These tests passed when using the
fedora client's IT suite's Fedora and elasticsearch servers, but the ITs
did not pass when using the current deposit service's IT suite's
elasticsearch server.  Until this is resolved, these changes should not
be merged.